Make page_count and current accessible from Twig
There are times when it is necessary to show the page_count and current page in Twig templates (for example, "2/6" to mean "page 2 of 6"). This pull request aims to expose variables `current`, `items_per_page`, and `page_count`.

I am not good with making pull requests, so if there is any mistake, please tell me, and I will try my best to fix it. Thanks!